requirements.txt
attrs==23.1.0
certifi==2023.7.22
chromedriver-binary-auto==0.3.1
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
natsort==8.2.0
outcome==1.2.0
pymongo==4.5.0
PySocks==1.7.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.16
wsproto==1.2.0
